Move EventCard to global components cleanup
