+++ frontend/src/components/EventCard.tsx
